requirements.txt
Pillow==9.1.1
streamlit==1.10.0
torch==1.11.0
torchvision==0.12.0
